refactor(fs): extract convertError into error-rebrand.ts
